requirements.txt
asgiref==3.8.1
cffi==1.17.0
Django==5.0.6
gunicorn==23.0.0
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.22
pyheif==0.7.1
sqlparse==0.5.0
